Run replace for addToMany
fixes #279
